A bunch of changes, mostly involving scoring and histograms.

+++ src/ScoreKeeper7K_rank.cpp
@@ -1,15 +1,16 @@
 #include "Global.h"
 #include "ScoreKeeper.h"
 
+
 void ScoreKeeper7K::update_ranks(int ms){
 
 	if (ms < judgement_time[SKJ_W1]) ++rank_w1_count;
 	if (ms < judgement_time[SKJ_W2]) ++rank_w2_count;
 	if (ms < judgement_time[SKJ_W3]) ++rank_w3_count;
 
-	int rank_w1_pts = max(rank_w1_count * 2 - total_notes, 0);
-	int rank_w2_pts = max(rank_w2_count * 2 - total_notes, 0);
-	int rank_w3_pts = max(rank_w3_count * 2 - total_notes, 0);
+	long long rank_w1_pts = max(rank_w1_count * 2 - total_notes, 0LL);
+	long long rank_w2_pts = max(rank_w2_count * 2 - total_notes, 0LL);
+	long long rank_w3_pts = max(rank_w3_count * 2 - total_notes, 0LL);
 
 	rank_pts = rank_w1_pts + rank_w2_pts + rank_w3_pts;
